server/story_tts_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-dotenv>=1.0.0
torch==2.2.2
torchaudio==2.2.2
transformers==4.33.3
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0
TTS==0.22.0
numpy==1.26.4
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
inflect==7.0.0
peft>=0.4.0
bitsandbytes>=0.39.0
